Add files via upload
Repository Files Description
1.Data270_project_Individual_Updated.ipynb

This Jupyter notebook contains the implementation of an individual XGBoost model for the H-Predict project. It includes both the baseline model and the hyperparameter-tuned model, demonstrating the process of model training, validation, and testing on healthcare cost prediction data.
2.xgb_model_data.pkl

A pickle file containing the trained baseline XGBoost model along with predictions and performance metrics (MSE, RMSE, R²) for training, validation, and testing datasets. This file facilitates easy loading of the model for further analysis or validation without the need to retrain.
3.xgb_model_results_hyperparameter.pkl

A pickle file containing the hyperparameter-tuned XGBoost model from the H-Predict project. This includes the optimized model parameters along with predictions and performance metrics (MSE, RMSE, R²) across training, validation, and testing datasets. It's used to assess the improvement in prediction accuracy due to hyperparameter tuning.
4.GWAR_Reference_ccs_pcode.csv

A CSV file that contains procedural codes relevant to the GWAR dataset, used as a reference for mapping or analysis purposes in healthcare cost prediction.
5.GWAR_Reference_ccs_dcode.csv

A CSV file that includes diagnostic codes from the GWAR dataset. This reference file is crucial for feature engineering and understanding the context of healthcare treatments in cost predictions.

6. Please find the Base data file and the final overall python file with all the files in this sharepoint link:
https://sjsu0-my.sharepoint.com/:f:/g/personal/kavana_anil_sjsu_edu/EnUZzFoPOTJAimwFJYeTUXABNThxcgeui0HdaOl0VQqBNA?e=LG3Cnd